# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/group/:id": {
            "delete": {
                "produces": [
                    "application/json"
                ],
                "summary": "删除群聊",
                "parameters": [
                    {
                        "description": "房间ID",
                        "name": "id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/resp.ResponseData"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/resp.ResponseData"
                        }
                    }
                }
            }
        },
...
        "/api/group/getGroupMemberList": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "退出群聊",
                "parameters": [
                    {
                        "description": "房间id",
                        "name": "id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/resp.ResponseData"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/resp.ResponseData"
                        }
                    }
                }
            }
        },
        "/api/group/grantAdministrator": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "授予管理员权限",
                "parameters": [
                    {
                        "description": "房间id",
                        "name": "room_id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "授权用户id",
                        "name": "grant_uid",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/resp.ResponseData"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/resp.ResponseData"
                        }
                    }
                }
            }
        },
        "/api/group/join": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "加入群聊",
                "parameters": [
                    {
                        "description": "房间id",
                        "name": "id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/resp.ResponseData"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/resp.ResponseData"
                        }
                    }
                }
            }
        },
        "/api/group/quit": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "退出群聊",
                "parameters": [
                    {
                        "description": "房间id",
                        "name": "id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/resp.ResponseData"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/resp.ResponseData"
                        }
                    }
                }
            }
        },
        "/api/group/removeAdministrator": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "移除管理员权限",
                "parameters": [
                    {
                        "description": "房间id",
                        "name": "room_id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "授权用户id",
                        "name": "remove_uid",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/resp.ResponseData"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/resp.ResponseData"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "resp.ResponseData": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "响应数据"
                },
                "message": {
                    "description": "响应消息",
                    "type": "string"
                },
                "success": {
                    "description": "是否成功",
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}